feat: pagina comunicação atualizada com botao marcar todas e entradas corretas
